dev/sdk_tools/pyproject.toml
[project]
name = "arva_dev_tools"
version = "dev-.1"
dependencies = [
    "pyyaml==6.0.2",
    "pathlib==1.0.1"
]